Add the e2controller, a way to interact with tools in e2.
Adds e2 controller functions and the E2 Controller tool. This is a huge update, it allows for the tool users to select an e2 chip either by clicking on a chip manually with the tool or by using the e2 chip to set their tool to select it. It can receive the trace/ranger data that a tool gets in left click / right click / reload events, and even is triggered when someone selects the chip. (Only when the selected chip changes, so no infinite loop)