vex library rewrite, webMaterials rewrite
Very cool rewrite of the VExtensions library. It exists in the SHARED realm now. This should simplify a ton of stuff. Also, more or less finalized webMaterials. I need to look into some problems with setMaterial with them, because it seems to be inconsistent and sometimes doesn't even set the material.