Fixed a lot of vex functions
Turned guessE2Type and sanitizeLuaVar into one local function, getVarTypeAndSanitize to optimize it, this function returns typeid and sanitized var if there is a need to sanitize it. There is still much to do, this should be the last thing in this PR. This shouldn't break anything that you added. Also, array type inferrence is more accurate since it looks for whether the table contains another table in it.