[WIP] The Patch PR (#15)
* Initial gitignore file

* Make sure git ignores these 2 folders (which are created by the editors/IDE), and we don't want these to be in our repository.

* Let The Patch begin 💥

* Make some changes to sv_vex_main

I fixed a typo in rangerSetFilter that didn't use fixed instead of filter which would've completely defeated the purpose of the loop.
Changed the hook name of hidechatply, super outdated :/
Made table_insert a local var, since we'll be spamming that in rangerSetFilter.
Also cleans table of rangerOffsetManual, and made a local function.

* Review for sv_coroutines and rune2

Coroutines:
Yes, we do set the value to nil, idk why I didn't here

Ln 150: We have to check if prfDataOrDone is true as a bool otherwise it'd constantly return true and never actually finish the function

RunE2:
Set the cost of try to 5 ops
Got rid of catch() as a whole

* Make `defined` function prefer the builtin instead of UDF.

* Implement `udfAll` , partially complete.

* Add `getExtensionsInfo` E2 function

* Add `getConstants` E2 function

* Change `vex.getE2Func`, allowing it to return a table

* Removed `udfNames` function

* Renamed `udfAll` to `getUserFunctionInfo`, and constants too.

* Refactor luaTableToE2, getE2Type

Shouldn't have to do anything else. You could still localize some variables that are used a ton but I don't think it's necessary
Added vex.guessE2Type and vex.sanitizeLuaVar, which are just helpers for getE2Type and luaTableToE2.
Got rid of all of the sanitization in luaTableToE2, we don't need to sanitize anything in there anyways besides bool values.

Will fix printGlobal to use the new type system soon

* Add `getBuiltinFuncInfo` E2 function

* Update printGlobal to use new vex.getE2Type return

Since vex.getE2Type now returns the type id rather than the name / some random thing like "VECTOR", we will use that here

Also fixes it using the old convar name without the vex prefix... I swear I fixed this ages ago... 🤷‍♂️

* Oh no... these should not be global variables.

* arrayOptimization is back

Was gone for a bit, my bad
Also got rid of some useless variables and actually made it recursive..

* Add `getTypeInfo` E2 function

* Add the missing local

* Import into locals

* Pass arrayOptimization recursively

* Bring backward compatibility (getE2Type)

* And now commit the rest.

* Import locals.

* Ignore the leading `@` for `getFunctionPath` function

* E1 normal to E2 number type alias fix

* Localization

* Fix helper functions.

* Hotfix `guessE2Type`, just let table go through

* Whoops, remove unused `true` literal

* Fix typos

* Specialized `luaTableToE2` for performance - just for SelfAware2 ✨

* Remove unused local

* Make the specialized `luaTableToE2` even shorter

* Just a couple of final touches

* Fixed a lot of vex functions

Turned guessE2Type and sanitizeLuaVar into one local function, getVarTypeAndSanitize to optimize it, this function returns typeid and sanitized var if there is a need to sanitize it. There is still much to do, this should be the last thing in this PR. This shouldn't break anything that you added. Also, array type inferrence is more accurate since it looks for whether the table contains another table in it.

* Optimized and Updated sv_printGlobal

This is the only thing that depended on vex.getE2Type. Also localized the functions used several times and got rid of that type() spam...

* Get rid of unused locals

We stopped using table.IsEmpty since it'd have barely any impact on performance. + Got rid of things like isSequential and table.Copy since we're not using them anymore

* Breaking change: `rangerOffsetManual` returning `ranger` instead

* Remove unused locals

* Fix the broken `cleanupTable`

* Added rangerdata to the inferrence list

Also removed buildBody from rune2, and a bunch of other things

Co-authored-by: Vurv <56230599+Vurv78@users.noreply.github.com>
Co-authored-by: Vurv78 <Vurv78@users.noreply.github.com>