Change some things under the hood in sv_coroutines.lua
Made respective functions that allow for resuming / rebooting / yielding coroutines with optional data to pass to the e2, to avoid repeated code in the e2functions that we previously had. Also, added a new e2helperfunc "throw" which is literally just the lua function error() except it is in a string.format form and defaults to error at scope 0. Lowered the threshold for e2s to trigger the coroutine stack overflow, since there aren't really usecases for going past 39 recursions of creating coroutines...

I ended up getting rid of the "trim" functions since this was really intensive and a single StartsWith is fine for this case. If it is ever needed, bless github for keeping all of this code in history.

Still haven't addressed #18, that would need a 100% rewrite, but I think the code is more organized enough to start on that now.